refactor(useIsMounted): move from src/devtools to src/react and reuse the hook where mountedRef scenarios appears
