+++ src/react/useIsMutating.ts
@@ -4,6 +4,7 @@ import { notifyManager } from '../core/notifyManager'
 import { QueryKey } from '../core/types'
 import { MutationFilters, parseMutationFilterArgs } from '../core/utils'
 import { useQueryClient } from './QueryClientProvider'
+import { useIsMounted } from './utils'
 
 export function useIsMutating(filters?: MutationFilters): number
 export function useIsMutating(
@@ -14,7 +15,7 @@ export function useIsMutating(
   arg1?: QueryKey | MutationFilters,
   arg2?: MutationFilters
 ): number {
-  const mountedRef = React.useRef(false)
+  const isMounted = useIsMounted()
   const filters = parseMutationFilterArgs(arg1, arg2)
 
   const queryClient = useQueryClient()
@@ -29,11 +30,9 @@ export function useIsMutating(
   isMutatingRef.current = isMutating
 
   React.useEffect(() => {
-    mountedRef.current = true
-
     const unsubscribe = queryClient.getMutationCache().subscribe(
       notifyManager.batchCalls(() => {
-        if (mountedRef.current) {
+        if (isMounted()) {
           const newIsMutating = queryClient.isMutating(filtersRef.current)
           if (isMutatingRef.current !== newIsMutating) {
             setIsMutating(newIsMutating)
@@ -43,10 +42,9 @@ export function useIsMutating(
     )
 
     return () => {
-      mountedRef.current = false
       unsubscribe()
     }
-  }, [queryClient])
+  }, [queryClient, isMounted])
 
   return isMutating
 }
